refactor: rename ProductRepository interface package and update imports in ListProductsUseCase
